FINALLY figured out what was wrong with zexdoc!
The issue was prefixed instructions that dont map to anything..

The failing zexdoc test: `ld <bcdexya>, <bcdexya>` starts with the base
case 0xDD 0x40. This isn't a legit dd-prefixed instruction.

So in a real z80, it is ignored and just acts as 0x40. The emulation
was instead refetching the next opcode from the memory (the byte AFTER
0x40) and running that instead of just re-decoding the exsiting 0x40
opcode.

:-)